before refactored to polymorphism
